Use consistent format, with soft links, for versioned js downloads
