+++ _layouts/default.html
@@ -39,8 +39,7 @@
 
     <script type="text/javascript" src={{ "/js/jquery.js" | prepend: site.baseurl }}></script>
     <script src={{ "/bootstrap/js/bootstrap.min.js" | prepend: site.baseurl }} type="text/javascript"></script>
-    <script src={{ "/js/jquery-cookie-1.4.1/jquery.cookie.js" | prepend: site.baseurl }} type="text/javascript"></script>
-
+    <script src={{ "/js/jquery-cookie.js" | prepend: site.baseurl }} type="text/javascript"></script>-
     {% include_cached google_analytics.html %}
     <script type="text/javascript" src={{ "/js/toc.js" | prepend: site.baseurl }}></script>
