+++ _layouts/search.html
@@ -169,7 +169,7 @@ layout: default
 </script>
 {% endraw %}
 
-<script src="{{site.baseurl}}/js/lunr.min.js" type="text/javascript" charset="utf-8"></script>
+<script src="{{site.baseurl}}/js/lunr.js" type="text/javascript" charset="utf-8"></script>
 <script src="{{site.baseurl}}/js/mustache.js" type="text/javascript" charset="utf-8"></script>
 <script src="{{site.baseurl}}/js/pagination.js" type="text/javascript" charset="utf-8"></script>
 <!--<script src="/js/date.format.js" type="text/javascript" charset="utf-8"></script>-->
