+++ _layouts/stats.html
@@ -77,7 +77,7 @@ layout: default
 <script src="https://d3js.org/d3.v2.min.js"></script>
 <script src="{{site.baseurl}}/js/venn.js/venn.js"></script>
 <script src="{{site.baseurl}}/js/mds.js"></script>
-<script src="{{site.baseurl}}/js/numeric-1.2.6.js"></script>
+<script src="{{site.baseurl}}/js/numeric.js"></script>
 <script>
 // define sets and set set intersections
 var sets = [
