+++ js/jquery-cookie.js
@@ -1,0 +1,1 @@
+jquery-cookie-1.4.1.js
